+++ visualization/app/codeCharta/ui/export3DMapButton/export3DMapButton.module.ts
@@ -2,10 +2,16 @@ import { CommonModule } from "@angular/common"
 import { NgModule } from "@angular/core"
 import { Export3DMapButtonComponent } from "./export3DMapButton.component"
 import { ActionIconModule } from "../actionIcon/actionIcon.module"
+import { HttpClientModule } from "@angular/common/http"
+import { MaterialModule } from "../../../material/material.module"
+import { Export3DMapDialogComponent } from "./export3DMapDialog/export3DMapDialog.component"
+import { MatSliderModule } from "@angular/material/slider"
+import { FormsModule } from "@angular/forms"
+import { LabelledColorPickerModule } from "../labelledColorPicker/labelledColorPicker.module"
 
 @NgModule({
-    imports: [CommonModule, ActionIconModule],
-    declarations: [Export3DMapButtonComponent],
+    imports: [CommonModule, ActionIconModule, HttpClientModule, MaterialModule, MatSliderModule, FormsModule, LabelledColorPickerModule],
+    declarations: [Export3DMapButtonComponent, Export3DMapDialogComponent],
     exports: [Export3DMapButtonComponent]
 })
 export class Export3DMapButtonModule {}